+++ CHANGELOG.md
@@ -7,18 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking changes
+
+- fix(k8sprocessor)!: deprecate clusterName setting [#578]
+
 ### Added
 
 - feat: enable rawk8seventsreceiver [#576]
-
-[Unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.50.0-sumo-0...main
-[576]: https://github.com/SumoLogic/sumologic-otel-collector/pull/576
 
 ### Fixed
 
 - fix: fix(cascadingfilterprocessor): do not attach sampling.rule attribute if trace accept rules are not specified [#575][#575]
 
+[Unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.50.0-sumo-0...main
+[#576]: https://github.com/SumoLogic/sumologic-otel-collector/pull/576
 [#575]: https://github.com/SumoLogic/sumologic-otel-collector/pull/575
+[#578]: https://github.com/SumoLogic/sumologic-otel-collector/pull/578
 
 ## [v0.50.0-sumo-0]